requirements.txt
aiohttp==3.12.15
websockets==12.0
PyYAML==6.0.2
python-dotenv==1.0.1
curl_cffi==0.6.0
